+++ packages/x-adapter-platform/src/mappers/requests/index.ts
@@ -3,6 +3,7 @@ export * from './next-queries-request.mapper';
 export * from './popular-searches-request.mapper';
 export * from './query-suggestions-request.mapper';
 export * from './recommendations-request.mapper';
+export * from './related-prompts-request.mapper';
 export * from './related-tags-request.mapper';
 export * from './search-request.mapper';
 export * from './semantic-queries-request.mapper';