+++ packages/x-adapter-platform/src/mappers/responses/index.ts
@@ -3,6 +3,7 @@ export * from './next-queries-response.mapper';
 export * from './query-suggestions-response.mapper';
 export * from './popular-searches-response.mapper';
 export * from './recommendations-response.mapper';
+export * from './related-prompts-response.mapper';
 export * from './related-tags-response.mapper';
 export * from './search-response.mapper';
 export * from './semantic-queries-response.mapper';